backend/requirement.txt
flask
flask-cors
python-dotenv
mysql-connector-python
tensorflow-cpu==2.16.1
numpy
pillow
requests
